moved tensorflow-specific stuff out of the evaluator
