+++ python/ray/rllib/ddpg/ddpg_evaluator.py
@@ -13,7 +13,6 @@ from ray.rllib.utils.sampler import SyncSampler
 
 
 import numpy as np
-import tensorflow as tf
 
 
 class DDPGEvaluator(PolicyEvaluator):
@@ -24,47 +23,15 @@ class DDPGEvaluator(PolicyEvaluator):
             registry, env_creator(config["env_config"]))
         self.config = config
 
-        self.sess = tf.Session()
-
-        with tf.variable_scope("model"):
-            self.model = DDPGModel(self.registry,
-                                   self.env,
-                                   self.config,
-                                   self.sess)
-        with tf.variable_scope("target_model"):
-            self.target_model = DDPGModel(self.registry,
-                                          self.env,
-                                          self.config,
-                                          self.sess)
-
-        self.setup_gradients()
-        self._setup_target_updates()
-
-        self.initialize()
-
-        # Set initial target weights to match model weights.
-        a_updates = []
-        for var, target_var in zip(self.model.actor_var_list,
-                                   self.target_model.actor_var_list):
-            a_updates.append(tf.assign(target_var, var))
-        actor_updates = tf.group(*a_updates)
-
-        c_updates = []
-        for var, target_var in zip(self.model.critic_var_list,
-                                   self.target_model.critic_var_list):
-            c_updates.append(tf.assign(target_var, var))
-        critic_updates = tf.group(*c_updates)
-        self.sess.run([actor_updates, critic_updates])
+        # contains model, target_model
+        self.model = DDPGModel(registry, self.env, config)
 
         self.sampler = SyncSampler(
-                        self.env, self.model, NoFilter(),
+                        self.env, self.model.model, NoFilter(),
                         config["num_local_steps"], horizon=config["horizon"])
 
         self.episode_rewards = [0.0]
         self.episode_lengths = [0.0]
-
-    def initialize(self):
-        self.sess.run(tf.global_variables_initializer())
 
     def sample(self):
         """Returns a batch of samples."""
@@ -95,96 +62,17 @@ class DDPGEvaluator(PolicyEvaluator):
             "num_episodes": len(self.episode_rewards),
         }
 
-    def _setup_target_updates(self):
-        """Set up target actor and critic updates."""
-        a_updates = []
-        for var, target_var in zip(self.model.actor_var_list,
-                                   self.target_model.actor_var_list):
-            a_updates.append(
-                tf.assign(target_var,
-                          self.config["tau"] * var
-                          + (1. - self.config["tau"]) * target_var)
-            )
-        actor_updates = tf.group(*a_updates)
-
-        c_updates = []
-        for var, target_var in zip(self.model.critic_var_list,
-                                   self.target_model.critic_var_list):
-            c_updates.append(
-                tf.assign(target_var,
-                          self.config["tau"] * var
-                          + (1. - self.config["tau"]) * target_var)
-            )
-        critic_updates = tf.group(*c_updates)
-        self.target_updates = [actor_updates, critic_updates]
-
     def update_target(self):
         """Updates target critic and target actor."""
-        self.sess.run(self.target_updates)
-
-    def setup_gradients(self):
-        """Setups critic and actor gradients."""
-        self.critic_grads = tf.gradients(self.model.critic_loss,
-                                         self.model.critic_var_list)
-        c_grads_and_vars = list(zip(self.critic_grads,
-                                self.model.critic_var_list))
-        c_opt = tf.train.AdamOptimizer(self.config["critic_lr"])
-        self._apply_c_gradients = c_opt.apply_gradients(c_grads_and_vars)
-
-        self.actor_grads = tf.gradients(-self.model.cn_for_loss,
-                                        self.model.actor_var_list)
-        a_grads_and_vars = list(zip(self.actor_grads,
-                                self.model.actor_var_list))
-        a_opt = tf.train.AdamOptimizer(self.config["actor_lr"])
-        self._apply_a_gradients = a_opt.apply_gradients(a_grads_and_vars)
+        self.model.update_target()
 
     def compute_gradients(self, samples):
-        """ Returns gradient w.r.t. samples."""
-        # actor gradients
-        actor_actions = self.sess.run(self.model.output_action,
-                                      feed_dict={
-                                        self.model.obs: samples["obs"]
-                                      })
-
-        actor_feed_dict = {
-            self.model.obs: samples["obs"],
-            self.model.output_action: actor_actions,
-        }
-        self.actor_grads = [g for g in self.actor_grads if g is not None]
-        actor_grad = self.sess.run(self.actor_grads, feed_dict=actor_feed_dict)
-
-        # feed samples into target actor
-        target_Q_act = self.sess.run(self.target_model.output_action,
-                                     feed_dict={
-                                      self.target_model.obs: samples["new_obs"]
-                                     })
-        target_Q_dict = {
-            self.target_model.obs: samples["new_obs"],
-            self.target_model.act: target_Q_act,
-        }
-
-        target_Q = self.sess.run(self.target_model.critic_eval,
-                                 feed_dict=target_Q_dict)
-
-        # critic gradients
-        critic_feed_dict = {
-            self.model.obs: samples["obs"],
-            self.model.act: samples["actions"],
-            self.model.reward: samples["rewards"],
-            self.model.target_Q: target_Q,
-        }
-        self.critic_grads = [g for g in self.critic_grads if g is not None]
-        critic_grad = self.sess.run(self.critic_grads,
-                                    feed_dict=critic_feed_dict)
-        return (critic_grad, actor_grad), {}
+        """Returns critic, actor gradients."""
+        return self.model.compute_gradients(samples)
 
     def apply_gradients(self, grads):
         """Applies gradients to evaluator weights."""
-        c_grads, a_grads = grads
-        critic_feed_dict = dict(zip(self.critic_grads, c_grads))
-        self.sess.run(self._apply_c_gradients, feed_dict=critic_feed_dict)
-        actor_feed_dict = dict(zip(self.actor_grads, a_grads))
-        self.sess.run(self._apply_a_gradients, feed_dict=actor_feed_dict)
+        self.model.apply_gradients(grads)
 
     def compute_apply(self, samples):
         grads, _ = self.compute_gradients(samples)
