+++ python/ray/rllib/ddpg/models.py
@@ -8,8 +8,140 @@ import numpy as np
 import tensorflow as tf
 import tensorflow.contrib.slim as slim
 
-
 class DDPGModel():
+    def __init__(self, registry, env, config):
+        self.registry = registry
+        self.env = env
+        self.config = config
+        self.sess = tf.Session()
+
+        with tf.variable_scope("model"):
+            self.model = DDPGActorCritic(self.registry,
+                                         self.env,
+                                         self.config,
+                                         self.sess)
+        with tf.variable_scope("target_model"):
+            self.target_model = DDPGActorCritic(self.registry,
+                                                self.env,
+                                                self.config,
+                                                self.sess)
+        self._setup_gradients()
+        self._setup_target_updates()
+
+        self.initialize()
+        # Set initial target weights to match model weights.
+        self._initialize_target_weights()
+
+    def _setup_gradients(self):
+        """Setups critic and actor gradients."""
+        self.critic_grads = tf.gradients(self.model.critic_loss,
+                                         self.model.critic_var_list)
+        c_grads_and_vars = list(zip(self.critic_grads,
+                                self.model.critic_var_list))
+        c_opt = tf.train.AdamOptimizer(self.config["critic_lr"])
+        self._apply_c_gradients = c_opt.apply_gradients(c_grads_and_vars)
+
+        self.actor_grads = tf.gradients(-self.model.cn_for_loss,
+                                        self.model.actor_var_list)
+        a_grads_and_vars = list(zip(self.actor_grads,
+                                self.model.actor_var_list))
+        a_opt = tf.train.AdamOptimizer(self.config["actor_lr"])
+        self._apply_a_gradients = a_opt.apply_gradients(a_grads_and_vars)
+
+    def compute_gradients(self, samples):
+        """ Returns gradient w.r.t. samples."""
+        # actor gradients
+        actor_actions = self.sess.run(self.model.output_action,
+                                      feed_dict={
+                                        self.model.obs: samples["obs"]
+                                      })
+
+        actor_feed_dict = {
+            self.model.obs: samples["obs"],
+            self.model.output_action: actor_actions,
+        }
+        self.actor_grads = [g for g in self.actor_grads if g is not None]
+        actor_grad = self.sess.run(self.actor_grads, feed_dict=actor_feed_dict)
+
+        # feed samples into target actor
+        target_Q_act = self.sess.run(self.target_model.output_action,
+                                     feed_dict={
+                                      self.target_model.obs: samples["new_obs"]
+                                     })
+        target_Q_dict = {
+            self.target_model.obs: samples["new_obs"],
+            self.target_model.act: target_Q_act,
+        }
+
+        target_Q = self.sess.run(self.target_model.critic_eval,
+                                 feed_dict=target_Q_dict)
+
+        # critic gradients
+        critic_feed_dict = {
+            self.model.obs: samples["obs"],
+            self.model.act: samples["actions"],
+            self.model.reward: samples["rewards"],
+            self.model.target_Q: target_Q,
+        }
+        self.critic_grads = [g for g in self.critic_grads if g is not None]
+        critic_grad = self.sess.run(self.critic_grads,
+                                    feed_dict=critic_feed_dict)
+        return (critic_grad, actor_grad), {}
+
+    def apply_gradients(self, grads):
+        """Applies gradients to evaluator weights."""
+        c_grads, a_grads = grads
+        critic_feed_dict = dict(zip(self.critic_grads, c_grads))
+        self.sess.run(self._apply_c_gradients, feed_dict=critic_feed_dict)
+        actor_feed_dict = dict(zip(self.actor_grads, a_grads))
+        self.sess.run(self._apply_a_gradients, feed_dict=actor_feed_dict)
+
+    def _setup_target_updates(self):
+        """Set up target actor and critic updates."""
+        a_updates = []
+        for var, target_var in zip(self.model.actor_var_list,
+                                   self.target_model.actor_var_list):
+            a_updates.append(
+                tf.assign(target_var,
+                          self.config["tau"] * var
+                          + (1. - self.config["tau"]) * target_var)
+            )
+        actor_updates = tf.group(*a_updates)
+
+        c_updates = []
+        for var, target_var in zip(self.model.critic_var_list,
+                                   self.target_model.critic_var_list):
+            c_updates.append(
+                tf.assign(target_var,
+                          self.config["tau"] * var
+                          + (1. - self.config["tau"]) * target_var)
+            )
+        critic_updates = tf.group(*c_updates)
+        self.target_updates = [actor_updates, critic_updates]
+
+    def update_target(self):
+        """Updates target critic and target actor."""
+        self.sess.run(self.target_updates)
+
+    def initialize(self):
+        self.sess.run(tf.global_variables_initializer())
+
+    def _initialize_target_weights(self):
+        a_updates = []
+        for var, target_var in zip(self.model.actor_var_list,
+                                   self.target_model.actor_var_list):
+            a_updates.append(tf.assign(target_var, var))
+        actor_updates = tf.group(*a_updates)
+
+        c_updates = []
+        for var, target_var in zip(self.model.critic_var_list,
+                                   self.target_model.critic_var_list):
+            c_updates.append(tf.assign(target_var, var))
+        critic_updates = tf.group(*c_updates)
+        self.sess.run([actor_updates, critic_updates])
+
+
+class DDPGActorCritic():
     other_output = []
     is_recurrent = False
